Support fractional minute durations and time ranges
Change TrackSegment and ScheduledSegment durationMinutes from Int to Double to allow fractional minutes. Update ScheduleLoader to validate non-finite values and compute durations using fractional minutes. Enhance DurationFormatting to show seconds, multi-unit labels (hr/min/sec) and add a timeRange formatter. Update TimelineView to pass eventStartDate and display segment time ranges. Add Toraku.entitlements and wire CODE_SIGN_ENTITLEMENTS in the Xcode project. Add unit tests covering fractional durations and the new formatting/time-range behavior.

+++ Toraku/Toraku/Models/ScheduledSegment.swift
@@ -10,7 +10,7 @@ struct ScheduledSegment: Identifiable, Equatable {
         endOffset - startOffset
     }
 
-    var durationMinutes: Int {
+    var durationMinutes: Double {
         segment.durationMinutes
     }
 }
